+++ src/main/java/com/prography/yakgwa/domain/meet/controller/MeetApi.java
@@ -1,9 +1,6 @@
 package com.prography.yakgwa.domain.meet.controller;
 
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.prography.yakgwa.domain.meet.controller.req.CreateMeetRequest;
-import com.prography.yakgwa.domain.meet.controller.res.CreateMeetResponse;
 import com.prography.yakgwa.domain.meet.controller.res.MeetInfoWithParticipantResponse;
 import com.prography.yakgwa.domain.meet.controller.res.MeetWithStatusInfoResponse;
 import com.prography.yakgwa.domain.meet.controller.res.PostConfirmMeetInfoResponse;
@@ -13,13 +10,9 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
 
 @Tag(name = "Meet", description = "모임관련 API입니다.")
 public interface MeetApi {
-    @Operation(summary = "모임 생성 API", description = "")
-    SuccessResponse<CreateMeetResponse> create(@AuthenticationPrincipal CustomUserDetail user,
-                                               @RequestBody CreateMeetRequest createMeetRequest) throws JsonProcessingException;
 
     @Operation(summary = "모임의 상세정보 조회 API")
     SuccessResponse<MeetInfoWithParticipantResponse> findDetail(@PathVariable("meetId") Long meetId);
