+++ src/main/java/com/prography/yakgwa/domain/meet/service/req/MeetCreateRequestDto.java
@@ -62,7 +62,4 @@ public class MeetCreateRequestDto {
                 .title(title)
                 .build();
     }
-
-
-
 }
